UPDATE : change folder structure contents is not changed by dsko
