--- java_thu/0ExampleFolder/CrawlingServerWithSelenium/apology.txt
@@ -1,7 +1,0 @@
-미안해요...
-저는 파일을 지워서....
-다시 만들기 귀찮기도 하고...
-
-죄송합니다.
-
-여러분은 안 지우셨죠??
