--- java_thu/0ExampleFolder/ProgrammersSolvedCode/1.java
@@ -1,16 +1,0 @@
-import java.util.HashSet;
-import java.util.Set;
-
-class Solution {
-     public int[] solution(int[] numbers) {
-        Set<Integer> set = new HashSet<>();
-
-        for(int i=0; i<numbers.length; i++) {
-            for(int j=i+1; j<numbers.length; j++) {
-                set.add(numbers[i] + numbers[j]);
-            }
-        }
-
-        return set.stream().sorted().mapToInt(Integer::intValue).toArray();
-    }
-}
