--- java_thu/0ExampleFolder/ProgrammersSolvedCode/4.java
@@ -1,21 +1,0 @@
-import java.util.*;
-
-class Solution {
-    public String solution(String[] participant, String[] completion) {
-        Map<String, Integer> map = new HashMap<>();
-
-        for (String p : participant) {
-            map.put(p, map.getOrDefault(p, 0) + 1);
-        }
-        for (String c : completion) {
-            map.put(c, map.get(c) - 1);
-        }
-        for (Map.Entry<String, Integer> entry : map.entrySet()) {
-            if (entry.getValue() > 0) {
-                return entry.getKey();
-            }
-        }
-        String answer = "";
-        return answer;
-    }
-}
