+++ java_thu/NohSeongMin/CrawlingServerWithSelenium/KNUNotice.java
@@ -1,0 +1,10 @@
+package haedal.selenium_one;
+
+import lombok.Data;
+
+@Data
+public class KNUNotice {
+    private String title;
+    private String writer;
+    private String date;
+}
