--- java_thu/NohSeongMin/KNUNotice.java
@@ -1,10 +1,0 @@
-package haedal.selenium_one;
-
-import lombok.Data;
-
-@Data
-public class KNUNotice {
-    private String title;
-    private String writer;
-    private String date;
-}
